myProject/requirements.txt
asgiref==3.8.1
contourpy==1.2.1
cycler==0.12.1
Django==5.0.7
fonttools==4.53.1
kiwisolver==1.4.5
matplotlib==3.9.1
networkx==3.3
numpy==2.0.1
packaging==24.1
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
scipy==1.14.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
